master: agregado endpoint para obtener reporte de carreras con info de inscriptos y egresados

+++ EjercicioIntegrador3/src/main/java/org/example/ejerciciointegrador3/dto/ReporteDTO.java
@@ -2,6 +2,8 @@ package org.example.ejerciciointegrador3.dto;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @ToString
@@ -10,7 +12,7 @@ import lombok.*;
 public class ReporteDTO {
 
     private String carrera;
-    private Integer anio;
+    private Long anio;
     private Long inscriptos;
     private Long egresados;
